[foe] Catch exceptions thrown by the authorization function.
Summary:
This modifies foe's wrap-authentication function to catch exceptions
thrown by the chosen authorization function. This allows for some
slightly richer surfacing of potential authentication issues.

I'll be the first to acknowledge that this is a somewhat experimental
solution. Still, it's the best I've got after an hour of thinking about
it.

Test Plan: CircleCI

Reviewers: ctdean, jim, mike, keith

Reviewed By: keith

Subscribers: Lighthouse, jim

Differential Revision: http://phabricator.corp.standardtreasury.com/D3792